requirements-test.txt
# Test-only dependencies (run from the repo root)
# Install with: pip install -r requirements-test.txt
pytest>=8.0
PyNaCl>=1.5.0
keyring>=25.0
pyyaml>=6.0
